+++ packages/amplify-schema-validator/src/__tests__/schemas/invalid-use-directives-from-newer-transformer-version.graphql
@@ -1,11 +1,5 @@
 type Post {
     id: ID! @primaryKey
     title: String! @index
-    blog: Blog! @hasOne
-}
-
-type Blog {
-    id: ID!
-    name: String!
-    posts: [Post]! @hasMany
+    blog: String!
 }
