+++ packages/amplify-schema-validator/src/__tests__/validates-use-directives-from-newer-transformer-version.test.ts
@@ -1,5 +1,6 @@
-import { ValidateSchemaProps, validateSchemaWithContext } from '..';
+import { validateSchemaWithContext } from '..';
 import { readSchema } from './helpers/readSchema';
+import { ValidateSchemaProps } from '../helpers/schema-validator-props';
 
 const schemaProps: ValidateSchemaProps = {
   graphqlTransformerVersion: 1,
@@ -9,7 +10,7 @@ const schemaProps: ValidateSchemaProps = {
 describe('Validate Schema', () => {
   it('fails validation when schema uses new directives in old transformer version', () => {
     const schema = readSchema('invalid-use-directives-from-newer-transformer-version.graphql');
-    const errorRegex = 'InvalidDirectiveError - Your GraphQL Schema is using @primaryKey, @index, @hasOne, @hasMany directives from a newer version of the GraphQL Transformer. Visit https://docs.amplify.aws/cli/migration/transformer-migration/ to learn how to migrate your GraphQL schema.';
+    const errorRegex = 'InvalidDirectiveError - Your GraphQL Schema is using @primaryKey, @index directives from a newer version of the GraphQL Transformer. Visit https://docs.amplify.aws/cli/migration/transformer-migration/ to learn how to migrate your GraphQL schema.';
     expect(() => validateSchemaWithContext(schema, schemaProps)).toThrow(errorRegex);
   });
 });
