+++ packages/amplify-schema-validator/src/__tests__/validates-use-directives-from-older-transformer-version.test.ts
@@ -1,4 +1,5 @@
-import { ValidateSchemaProps, validateSchemaWithContext } from '..';
+import { validateSchemaWithContext } from '..';
+import { ValidateSchemaProps } from '../helpers/schema-validator-props';
 import { readSchema } from './helpers/readSchema';
 
 const schemaProps: ValidateSchemaProps = {
