+++ packages/amplify-schema-validator/src/helpers/schema-validator-props.ts
@@ -1,0 +1,4 @@
+export type ValidateSchemaProps = {
+    graphqlTransformerVersion: number;
+    isDataStoreEnabled: boolean;
+};
