+++ packages/amplify-schema-validator/src/validators/use-directives-from-newer-transformer-version.ts
@@ -3,7 +3,8 @@ import {
   Kind,
   ObjectTypeDefinitionNode,
 } from 'graphql';
-import { InvalidDirectiveError } from '../exceptions/invalid-directive-error';
+import { ValidateSchemaProps } from '../helpers/schema-validator-props';
+import { transformerValidationErrors } from '../helpers/transformer-validation';
 
 /**
  * Validates that @belongsTo, @connection directives from older graphql transformer version is not used in version 1.0
@@ -11,11 +12,6 @@ import { InvalidDirectiveError } from '../exceptions/invalid-directive-error';
  * @param schema graphql schema
  * @returns true if correct directives are used
  */
-
-type ValidateSchemaProps = {
-  graphqlTransformerVersion: number;
-  isDataStoreEnabled: boolean;
-};
 
 const GRAPHQL_TRANSFORMER_V2_DIRECTIVES = ['hasOne', 'index', 'primaryKey', 'belongsTo', 'manyToMany', 'hasMany', 'default'];
 export const validateDirectivesFromNewerTransformerVersionAreNotUsed = (
@@ -25,7 +21,6 @@ export const validateDirectivesFromNewerTransformerVersionAreNotUsed = (
     return [];
   }
 
-  const errors: Error[] = [];
   const objectTypeDefinitions = schema.definitions.filter(
     (defintion) => defintion.kind === Kind.OBJECT_TYPE_DEFINITION,
   ) as ObjectTypeDefinitionNode[];
@@ -47,15 +42,6 @@ export const validateDirectivesFromNewerTransformerVersionAreNotUsed = (
       });
     });
   });
-  if (v2DirectivesInUse.size > 0) {
-    const errorMessage = `Your GraphQL Schema is using ${Array.from(v2DirectivesInUse.values())
-      .map((directive) => `${directive}`)
-      .join(', ')} ${
-      v2DirectivesInUse.size > 1 ? 'directives' : 'directive'
-    } from a newer version of the GraphQL Transformer. Visit https://docs.amplify.aws/cli/migration/transformer-migration/ to learn how to migrate your GraphQL schema.`;
-    errors.push(new InvalidDirectiveError(
-      errorMessage,
-    ));
-  }
-  return errors;
+
+  return transformerValidationErrors(v2DirectivesInUse, props.graphqlTransformerVersion);
 };
