+++ packages/amplify-schema-validator/src/validators/use-directives-from-older-transformer-version.ts
@@ -3,7 +3,8 @@ import {
   Kind,
   ObjectTypeDefinitionNode,
 } from 'graphql';
-import { InvalidDirectiveError } from '../exceptions/invalid-directive-error';
+import { ValidateSchemaProps } from '../helpers/schema-validator-props';
+import { transformerValidationErrors } from '../helpers/transformer-validation';
 
 /**
  * Validates that @key, @connection directives from older graphql transformer version is not used in version 2.0
@@ -12,10 +13,6 @@ import { InvalidDirectiveError } from '../exceptions/invalid-directive-error';
  * @returns true if correct directives are used
  */
 
-type ValidateSchemaProps = {
-  graphqlTransformerVersion: number;
-  isDataStoreEnabled: boolean;
-};
 
 export const validateDirectivesFromOlderTransformerVersionAreNotUsed = (
   schema: DocumentNode, props: ValidateSchemaProps,
@@ -24,7 +21,6 @@ export const validateDirectivesFromOlderTransformerVersionAreNotUsed = (
     return [];
   }
 
-  const errors: Error[] = [];
   const objectTypeDefinitions = schema.definitions.filter(
     (defintion) => defintion.kind === Kind.OBJECT_TYPE_DEFINITION,
   ) as ObjectTypeDefinitionNode[];
@@ -45,15 +41,5 @@ export const validateDirectivesFromOlderTransformerVersionAreNotUsed = (
       v1DirectivesInUse.add('@connection');
     }
   });
-  if (v1DirectivesInUse.size > 0) {
-    const errorMessage = `Your GraphQL Schema is using ${Array.from(v1DirectivesInUse.values())
-      .map((directive) => `${directive}`)
-      .join(', ')} ${
-      v1DirectivesInUse.size > 1 ? 'directives' : 'directive'
-    } from an older version of the GraphQL Transformer. Visit https://docs.amplify.aws/cli/migration/transformer-migration/ to learn how to migrate your GraphQL schema.`;
-    errors.push(new InvalidDirectiveError(
-      errorMessage,
-    ));
-  }
-  return errors;
+  return transformerValidationErrors(v1DirectivesInUse, props.graphqlTransformerVersion);
 };
